Refactor NexusPHP adapters to use NexusPHPHelper mixin
- Created NexusPHPHelper mixin to consolidate duplicate logic
- Extracted `getDownLoadHash`, `_parseTagType`, and `_parseDiscountType` from `NexusPHPAdapter` and `NexusPHPWebAdapter`
- Mixed `NexusPHPHelper` into both adapters and replaced usages to reduce duplication and enhance maintainability
